+++ jest.config.ts
@@ -24,4 +24,13 @@ module.exports = {
   transform: {
     '^.+\\.tsx?$': 'ts-jest',
   },
+  moduleNameMapper: {
+    '\\.(css|scss)$': '<rootDir>/src/test/__mocks__/styleMock.ts',
+    '\\.svg$': '<rootDir>/src/test/__mocks__/fileMock.ts',
+    '@mui/material/styles/createPalette':
+      '<rootDir>/src/test/__mocks__/createPalette.ts',
+    '@mui/material/styles/createTypography':
+      '<rootDir>/src/test/__mocks__/createTypography.ts',
+  },
+  setupFiles: ['<rootDir>/src/test/setupTests.ts'],
 }