+++ src/components/basic/Dropzone/DropPreviewFile.stories.tsx
@@ -18,7 +18,7 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-import { ComponentStory } from '@storybook/react'
+import { type ComponentStory } from '@storybook/react'
 
 import { DropPreviewFile as Component } from './components/DropPreviewFile'
 import { UploadStatus } from './types'
@@ -53,7 +53,9 @@ export default {
   },
 }
 
-const Template: ComponentStory<typeof Component> = (args: any) => {
+const Template: ComponentStory<typeof Component> = (
+  args: React.ComponentProps<typeof Component>
+) => {
   const { name, size, status, progressPercent, ...props } = args
   return (
     <Component