+++ src/components/basic/UserAvatar/UserAvatar.stories.tsx
@@ -18,7 +18,7 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-import { ComponentStory } from '@storybook/react'
+import { type ComponentStory } from '@storybook/react'
 
 import { UserAvatar as Component } from '.'
 
@@ -29,12 +29,12 @@ export default {
   argTypes: {},
 }
 
-const Template: ComponentStory<typeof Component> = (args: any) => (
-  <Component {...args} />
-)
+const Template: ComponentStory<typeof Component> = (
+  args: React.ComponentProps<typeof Component>
+) => <Component {...args} />
 
-export const userAvatar = Template.bind({})
-userAvatar.args = {
+export const UserAvatar = Template.bind({})
+UserAvatar.args = {
   notificationCount: 0,
   isNotificationAlert: false,
 }